Rename ruleSet files to revision for consistency
