+++ packages/sdk/src/revisions/v1/index.ts
@@ -1,3 +1,3 @@
 export * from './publicKey';
+export * from './revision';
 export * from './rule';
-export * from './ruleSet';
